refactor(charts): move stacked charts to dedicated directories
Move StackedBarChart and StackedColumnChart from their parent chart
directories to their own dedicated subdirectories for better organization
and consistency with project structure.
