+++ lib/datavisualisations/StackedBarChart/StackedBarChart.vue
