+++ lib/datavisualisations/StackedColumnChart/StackedColumnChart.vue
